Move most of ERSource to LXeSource
