+++ tests/test_source.py
@@ -4,7 +4,7 @@ import pytest
 import tensorflow as tf
 
 import flamedisx as fd
-from flamedisx.source import quanta_types
+from flamedisx.er_nr_base import quanta_types
 
 
 def np_lookup_axis1(x, indices, fill_value=0):
